Thirrja e metodave per krijimin e Task aktivitetit
Eshte realizuar thirrja e metodave per krijimin e taskActivity kur:
krijohet/updatohet tasku,
krijohet/fshihet/updatohet checklista
krijohet/fshihet/updatohet checklistItem
assign/unassign taskMember

+++ backend/Controllers/TaskMemberController.cs
@@ -4,9 +4,11 @@ using backend.DTOs.TaskMember.Output;
 using backend.Interfaces;
 using backend.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace backend.Controllers;
+
 [Route("backend/TaskMembers")]
 [ApiController]
 public class TaskMemberController : ControllerBase
@@ -20,14 +22,19 @@ public class TaskMemberController : ControllerBase
     private readonly IBoardRepository _boardRepo;
     private readonly IWorkspaceRepository _workspaceRepo;
 
+    private readonly UserManager<User> _userManager;
+    private readonly ITaskActivityRepository _taskActivityRepo;
+
     public TaskMemberController(ITaskMemberRepository taskMemberRepo,
-                                IMapper mapper,
-                                ITaskRepository taskRepo,
-                                IUserRepository userRepo,
-                                IMembersRepository memberRepo,
-                                IListRepository listRepo,
-                                IBoardRepository boardRepo,
-                                IWorkspaceRepository workspaceRepo)
+        IMapper mapper,
+        ITaskRepository taskRepo,
+        IUserRepository userRepo,
+        IMembersRepository memberRepo,
+        IListRepository listRepo,
+        IBoardRepository boardRepo,
+        IWorkspaceRepository workspaceRepo,
+        UserManager<User> userManager,
+        ITaskActivityRepository taskActivityRepo)
     {
         _taskMemberRepo = taskMemberRepo;
         _mapper = mapper;
@@ -37,6 +44,9 @@ public class TaskMemberController : ControllerBase
         _listRepo = listRepo;
         _boardRepo = boardRepo;
         _workspaceRepo = workspaceRepo;
+
+        _userManager = userManager;
+        _taskActivityRepo = taskActivityRepo;
     }
 
     [Authorize(AuthenticationSchemes = "Bearer")]
@@ -57,7 +67,7 @@ public class TaskMemberController : ControllerBase
         }
         catch (Exception e)
         {
-            return StatusCode(500, "Internal Server Error!"+e.Message);
+            return StatusCode(500, "Internal Server Error!" + e.Message);
         }
     }
 
@@ -84,7 +94,7 @@ public class TaskMemberController : ControllerBase
         }
         catch (Exception e)
         {
-            return StatusCode(500, "Internal Server Error!"+e.Message);
+            return StatusCode(500, "Internal Server Error!" + e.Message);
         }
     }
 
@@ -121,7 +131,7 @@ public class TaskMemberController : ControllerBase
             {
                 return NotFound("Workspace Not Found!");
             }
-            
+
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
             var isMember = userId != null && await _memberRepo.IsAMember(userId, workspace.WorkspaceId);
@@ -142,7 +152,7 @@ public class TaskMemberController : ControllerBase
         }
         catch (Exception e)
         {
-            return StatusCode(500, "Internal Server Error!"+e.Message);
+            return StatusCode(500, "Internal Server Error!" + e.Message);
         }
     }
 
@@ -166,7 +176,7 @@ public class TaskMemberController : ControllerBase
             {
                 return NotFound("User Not Found!");
             }
-            
+
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
 
@@ -175,7 +185,7 @@ public class TaskMemberController : ControllerBase
             {
                 return NotFound("Task Not Found!");
             }
-            
+
             var list = await _listRepo.GetListByIdAsync(task.ListId);
             if (list == null)
             {
@@ -193,6 +203,7 @@ public class TaskMemberController : ControllerBase
             {
                 return NotFound("Workspace Not Found");
             }
+
             var isMember = userId != null && await _memberRepo.IsAMember(userId, workspace.WorkspaceId);
 
             var memberAddedIsMember = await _memberRepo.IsAMember(addTaskMemberDto.UserId, workspace.WorkspaceId);
@@ -202,14 +213,49 @@ public class TaskMemberController : ControllerBase
                 return NotFound("The User is not a member of this workspace!");
             }
 
-            var memberAddedAlreadyATaskMember = await _taskMemberRepo.IsATaskMember(addTaskMemberDto.UserId, addTaskMemberDto.TaskId);
+            var memberAddedAlreadyATaskMember =
+                await _taskMemberRepo.IsATaskMember(addTaskMemberDto.UserId, addTaskMemberDto.TaskId);
             if (memberAddedAlreadyATaskMember)
             {
                 return BadRequest("User is already a member in this task!");
             }
-            
+
+            var taskMember = await _userManager.FindByIdAsync(addTaskMemberDto.UserId);
+            if (taskMember == null)
+            {
+                return NotFound("User not found!");
+            }
             if (isMember || userTokenRole == "Admin")
             {
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+                
+
+                var taskActivity = new TaskActivity
+                {
+                    TaskId = task.TaskId,
+                    UserId = userId,
+                    ActionType = "Assigned",
+                    EntityName = taskMember.Email+" to task "+task.Title,
+                    ActionDate = DateTime.Now
+                };
+
+                await _taskActivityRepo.CreateTaskActivityAsync(taskActivity);
                 await _taskMemberRepo.AddTaskMemberAsync(addTaskMemberDto);
                 return StatusCode(200, "Member added to Task!");
             }
@@ -226,7 +272,8 @@ public class TaskMemberController : ControllerBase
 
     [Authorize(AuthenticationSchemes = "Bearer")]
     [HttpPut("UpdateTaskMember")]
-    public async Task<IActionResult> UpdateTaskMember(UpdateTaskMemberDto updateTaskMemberDto)
+
+public async Task<IActionResult> UpdateTaskMember(UpdateTaskMemberDto updateTaskMemberDto)
     {
         if (!ModelState.IsValid)
         {
@@ -341,10 +388,46 @@ public class TaskMemberController : ControllerBase
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
             var isMember = await _memberRepo.IsAMember(userId, workspace.WorkspaceId);
             var isTaskMember = await _taskMemberRepo.IsATaskMember(removeTaskMemberDto.UserId, removeTaskMemberDto.TaskId);
+            
+            var taskMember = await _userManager.FindByIdAsync(removeTaskMemberDto.UserId);
+            if (taskMember == null)
+            {
+                return NotFound("User not found!");
+            }
             if (isMember || userTokenRole == "Admin")
             {
                 if (isTaskMember)
                 {
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    
+                    var taskActivity = new TaskActivity
+                    {
+                        TaskId = task.TaskId,
+                        UserId = userId,
+                        ActionType = "Assigned",
+                        EntityName = taskMember.Email+" to task "+task.Title,
+                        ActionDate = DateTime.Now
+                    };
+
+                    await _taskActivityRepo.CreateTaskActivityAsync(taskActivity);
                     await _taskMemberRepo.RemoveTaskMemberAsync(removeTaskMemberDto.TaskId, removeTaskMemberDto.UserId);
 
                     return Ok("Member removed from Task");
